Main - Creating custom Dao exception and message class to Author Dao methods. Creation save methods.

+++ src/main/java/br/edu/femass/daos/Dao.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface Dao<T> {
     void save(T t) throws Exception;
     List<T> getAll() throws Exception;
-    T getByCode(Integer code) throws Exception;
+    T getByCode(Long code) throws Exception;
     Long getNextCode() throws Exception;
     void update(T t) throws Exception;
     void delete(Long code) throws Exception;
